+++ bin/check_sample_sheet.py
@@ -11,26 +11,31 @@ def main():
     parser = argparse.ArgumentParser()
     parser.add_argument('sample_sheet')
     parser.add_argument('output')
+    parser.add_argument('approx_size')
     args = parser.parse_args()
 
     try:
         samples = pd.read_csv(args.sample_sheet, sep=None)
         if 'alias' in samples.columns:
-            if 'sample_name' in samples.columns:
+            if 'sample_id' in samples.columns:
                 sys.stderr.write(
                     "Warning: sample sheet contains both 'alias' and "
-                    'sample_name, using the former.')
-            samples['sample_name'] = samples['alias']
-        if 'barcode' not in samples.columns \
-                or 'sample_name' not in samples.columns:
+                    'sample_id, using the former.')
+            samples['sample_id'] = samples['alias']
+        if not set(['sample_id', 'barcode']).intersection(samples.columns):
             raise IOError()
     except Exception:
         raise IOError(
             "Could not parse sample sheet, it must contain two columns "
-            "named 'barcode' and 'sample_name' or 'alias'.")
+            "named 'barcode' and 'sample_id' or 'alias'.")
+    if 'approx_size' not in samples.columns:
+        print(
+            "approx_size column not found in sample_sheet"
+            "so using param.approx_size for all samples")
+        samples['approx_size'] = args.approx_size
     # check duplicates
     dup_bc = samples['barcode'].duplicated()
-    dup_sample = samples['sample_name'].duplicated()
+    dup_sample = samples['sample_id'].duplicated()
     if any(dup_bc) or any(dup_sample):
         raise IOError(
             "Sample sheet contains duplicate values.")