+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.7]
+### Changed
+- Moved bioinformatics from report to seperate processes
+### Added
+- Ability to define approx_size of sequence per sample in sample_sheet
+- Insert length to table
+- Output annotation bed files per sample
+
 ## [v0.1.6]
 ### Changed
 - Update schema for epi2melabs compatibility